pyproject.toml
[project]
name = "procurement-compliance-openenv"
version = "0.1.0"
description = "OpenEnv environment for procurement compliance review"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "websockets"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"